Test app minor update
- delay field type is changed from long to int
- NonNull annotations

+++ testapp/src/main/java/com/testapp/presenter/MainState.java
@@ -11,7 +11,7 @@ import java.util.Locale;
 public class MainState extends MvpState {
     public List<Event> events = new ArrayList<>();
     public boolean isEventAdded;
-    public long delay = 100;
+    public int delay = 100;
     public String text = "";
     public ActionDuration duration = ActionDuration.LongDuration;
     public boolean isSubscribedToConnectivity;
@@ -86,7 +86,7 @@ public class MainState extends MvpState {
         this.duration = duration;
     }
 
-    public void setDelay(long delay) {
+    public void setDelay(int delay) {
         setChanged(this.delay != delay);
         this.delay = delay;
     }
